requirements.txt
PyQt5
python-dotenv
langchain-core
langchain-openai
langchain-tools
langgraph